Add basic support for confluencebuilder
This is pretty much just a quick proof of concept. Also I am still
learning how to properly hook into the various extensions. Reading the
original source for this one did inform me on how to extend with custom
node types for specfic builders. Very nice that the vistors are part of
the extension vs the builder.